:recycle: DOP-4358 pull out some functions into its own utils to be reused, along with creating a currentUserCleanup function
